+++ MiniProjektPersistens/src/model/Equipment.java
@@ -6,7 +6,7 @@ public class Equipment extends Product {
 	private String type;
 	private String description;
 	
-	public Equipment(String type, String description, int productID, String name, double purchasePrice, double salePrice, double rentPrice, String productType, int supplierID) {
+	public Equipment(String type, String description, int productID, String name, float purchasePrice, float salePrice, float rentPrice, String productType, int supplierID) {
 		super(productID, name, purchasePrice, salePrice, rentPrice, productType, supplierID);
 //		this.id = id; //TODO id skal stemme overens med product
 		this.type = type;
